Fixed Test Kitchen - For Real This Time
Remember that other commit where I said that I had finally got Test Kitchen working well that
was a lie. This commit gets it working now that I have the correct data in place. This was because
the recipe wrote out a template file that I had commented out while working with it.

This template file looks at the various attributes of the nodes. This content is going to be written
to a yaml file instead of the index html page that I am currently using.