Integration of the Portal Site Project
So the portal site is no longer going to be a static html page but a sinatra application.
This portal site needs the node information stored in a yaml file that it will read from.
The portal site is a ruby application so it will need its gems install and then it
have be run as a daemonized process. This is not that elegant of a solution but we
are not looking for a long running process. One that simply can execute a few commands.