Copying credentials to the Chef Portal
The best way that we found to get the credentials to the remote machine is to create a role with that information.
Then use that role within the portal site execution. This will provide new default value that machine the classroom
name and other values necessary to create the create instances.